% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arq.R
\name{arq}
\alias{arq}
\title{Autoregressive quantile model}
\usage{
arq(r, p, x, model = "s", sv = NULL)
}
\arguments{
\item{r}{Vector returns on an asset.}

\item{p}{Probability level of the quantile (scalar).}

\item{x}{Vector of observations for a realized measure.}

\item{model}{Type of model. "s" is default. See the details.}

\item{sv}{Vector of starting values. This is an optional argument.}
}
\value{
A list containing:
\describe{
\item{\code{estimates}}{ Vector of estimated parameters.}
\item{\code{value}}{ Minimized value of the loss function.}
\item{\code{q}}{ Vector of fitted quantiles.}
}
}
\description{
Fits an autoregressive quantile model with realized measures.
}
\details{
Parameters are estimated minimizing the quantile loss function. Optimization is performed with the \code{optim()} function
alternating "BFGS" and "Nelder-Mead" algorithms until convergence as in Engle and Manganelli (2003).

\code{model} currently allows to select one quantile models: "s". For a quantile \eqn{q^{p}} at level \eqn{p} the model is
\deqn{q^{p}_t = b_0 + b_1 q^{p}_{t-1} + b_2 x_{t-1}} where \eqn{x_{t-1}} is the lagged value of the realized measure.
}
\examples{
# Fit the ARQ model
p   <- 0.95
fit <- arq(sp500$r, p, sp500$rv)

}
\references{
Engle, R. F., and Manganelli, S. (2004). CAViaR: Conditional autoregressive value at risk by regression quantiles. \emph{Journal of Business & Economic Statistics}, 22(4), 367-381.
}
\seealso{
\code{\link{arq.forecast}}
}
\author{
Luca Trapin
}
