% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinEx-package.R
\docType{package}
\name{FinEX-package}
\alias{FinEX-package}
\title{Financial Econometrics with extreme value statistics}
\description{
The package provides different econometrics techniques based on extreme value statistics to compute dynamic conditional risk measures used in financial econometrics.
}
\seealso{
Useful links
\itemize{
\item \url{https://github.com/lucatrapin/FinEx}
}
}
\author{
\describe{
\item{Mantainer:}{Luca Trapin \code{<luca.trapin@unibo.it>}}
\item{Author:}{Marco Bee \code{<marco.bee@unitn.it>}}
}
}
