% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.forecast.R
\name{req.forecast}
\alias{req.forecast}
\title{Realized extreme quantile forecasts}
\usage{
req.forecast(fit, x, alpha)
}
\arguments{
\item{fit}{Output object of the \code{arq()} function.}

\item{x}{Scalar with the realized measures used in the forecast.}

\item{alpha}{Probability level for the computation of the risk measures.}
}
\value{
A matrix containing the VaR and ES forecasts at level alpha.
}
\description{
Computes one-day-ahead forecasats of Value at Risk (VaR) and Expected Shortfall (ES) for a fitted REQ model.
}
\references{
Bee, M., Dupuis, D. J., and Trapin, L. (2018). Realized extreme quantile: A joint model for conditional quantiles and measures of volatility with EVT refinements. \emph{Journal of Applied Econometrics}, 33(3), 398-415.
}
\author{
Luca Trapin
}
