% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{req}
\alias{req}
\title{Realized Extreme Quantile}
\usage{
req(r, p, x, alpha, pa, model = "s", sv = NULL)
}
\arguments{
\item{r}{Vector of returns on an asset.}

\item{p}{Probability level of the dynamic threshold.}

\item{x}{Vector of observations for a realized measure.}

\item{alpha}{Probability level of the VaR.}

\item{pa}{Probability level used to set the threshold in the Hill estimator.}

\item{model}{Type of model: "s" is default. See details in the \code{arq()} function.}

\item{sv}{Vector of starting values for the \code{arq()} model. This is an optional argument.}
}
\value{
A list containing:
\describe{
\item{\code{estimates}}{Parameters of the \code{arq()} model and the tail index.}
\item{\code{value}}{Minimized value of the loss function from the \code{arq()} model.}
\item{\code{p}}{Probability level of the dynamic threshold.}
\item{\code{zk}}{Threshold value of the quantile residuals used to estimate the tail index.}
\item{\code{q}}{Vector containing the dynamic threshold computed with the \code{arq()} model.}
\item{\code{var}}{Vector containing the VaR estimates.}
}
}
\description{
Computes Value at Risk (VaR) using a dynamic threshold estimated with the \code{arq()} function
and extreme value refinements as in Bee et al. (2018).
}
\details{
Let \eqn{q^{p}} be the estimated dynamic threshold with the \code{arq()} model, the tail index (\eqn{\xi}) is computed on the quantile residuals, \eqn{z^{p}=\frac{r}{q^{p}},} using the the Hill estimator,
\deqn{\widehat{\xi}=\frac{1}{k} \sum^k_{j=1}\log\left(\frac{z^{p}_{(j)}}{z^{p}_{(k)}}\right),} where \eqn{z^{p}_{(k)}} is the order statistics associated to the probability level \eqn{pa}.
The VaR is computed as \deqn{VaR^{\alpha}_t=q^{p}_t z^{p}_k \left(\frac{p}{\alpha}\right)^{\xi}}
}
\examples{
# Fit the REQ model
p <- 0.95
pa <- 0.975
alpha <- 0.99
fit <- req(sp500$r, p, sp500$rv, alpha, pa)

}
\references{
Bee, M., Dupuis, D. J., and Trapin, L. (2018). Realized extreme quantile: A joint model for conditional quantiles and measures of volatility with EVT refinements. \emph{Journal of Applied Econometrics}, 33(3), 398-415.
}
\seealso{
\code{\link{arq}}, \code{\link{req.forecast}}
}
\author{
Luca Trapin
}
