% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpot.fitted.R
\name{rpot.fitted}
\alias{rpot.fitted}
\title{Realized Peaks over Threshold fitted values}
\usage{
rpot.fitted(fit, alpha)
}
\arguments{
\item{fit}{Output object of the \code{rpot()} function.}

\item{alpha}{Probability level for the computation of the risk measures.}
}
\value{
A matrix containing the VaR and ES fitted values at level alpha.
}
\description{
Computes fitted Value at Risk (VaR) and Expected Shortfall (ES) for a fitted RPoT model.
}
\details{
The risk measures are computed as in Bee et al. (2019):
\deqn{VaR^{\alpha}=q+\frac{\sigma}{\xi}\left(\frac{\phi}{1-\alpha}^{\xi}-1\right)}
\deqn{ES^{\alpha}=\frac{VaR^{\alpha}}{1-\xi}+\frac{\sigma-q\xi}{1-\xi}}
}
\references{
Bee, M., Dupuis, D. J., and Trapin, L. (2019). Realized Peaks over Threshold: A Time-Varying Extreme Value Approach with High-Frequency-Based Measures. \emph{Journal of Financial Econometrics}, 17(2), 254-283.
}
\seealso{
\code{\link{rpot.forecast}}
}
\author{
Luca Trapin
}
