% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpot.forecast.R
\name{rpot.forecast}
\alias{rpot.forecast}
\title{Realized Peaks over Threshold forecasts}
\usage{
rpot.forecast(fit, x, alpha)
}
\arguments{
\item{fit}{Output object of the \code{rpot()} function.}

\item{x}{Vector of realized measures used in the forecast. The intercept must be included.}

\item{alpha}{Probability level for the computation of the risk measures.}
}
\value{
A list containing:
\describe{
\item{\code{parameters}}{Vector containing the forecasts of phi, sigma, and xi.}
\item{\code{var}}{Var forecast at level alpha.}
\item{\code{es}}{ES forecast at level alpha.}
}
}
\description{
Computes one-day-ahead forecasts of Value at Risk (VaR) and Expected Shortfall (ES) for a fitted RPoT model.
}
\details{
The risk measures are computed as in Bee et al. (2019):
\deqn{VaR^{\alpha}=q+\frac{\sigma}{\xi}\left(\frac{\phi}{1-\alpha}^{\xi}-1\right)}
\deqn{ES^{\alpha}=\frac{VaR^{\alpha}}{1-\xi}+\frac{\sigma-q\xi}{1-\xi}}
}
\references{
Bee, M., Dupuis, D. J., and Trapin, L. (2019). Realized Peaks over Threshold: A Time-Varying Extreme Value Approach with High-Frequency-Based Measures. \emph{Journal of Financial Econometrics}, 17(2), 254-283.
}
\author{
Luca Trapin
}
