% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Daily returns and 5-minute realized variance for the SP500.}
\format{
An object of class \code{data.frame} with 3744 rows and 3 columns.
}
\usage{
data(sp500)
}
\description{
Daily observations from 2000 to 2014 from the Oxford-Man Realized library (Heber et al., 2009). These time series were used in the
empirical analysis by Bee and Trapin (2018).
}
\examples{
data(sp500)
returns <- sp500$r
realized_variance <- sp500$rv
}
\references{
Heber, G., Lunde, A., Shephard, N., and Sheppard, K. (2009). \emph{Oxford-Man Institute’s realized library}, version 0.1.

Bee, M., and Trapin, L. (2018). Estimating and forecasting conditional risk measures with extreme value theory: A review. \emph{Risks}, 6(2), 45.
}
\keyword{datasets}
